Pause moved from double click on Time Display to shift click on Time Display. It was no fun having seek forward and reverse be unresponsive because of double clicking.
Pause could go back in the context menu for n00bz to find but I don't like going into the context menu for something as frequent as pause ... since it would be easy to mistakenly hit star or delete.

Fixed a bug where randomPlay in SoundEngine was expecting a string, and was instead getting an event (when a song finished playing).